gmail-mcp-server/pyproject.toml
[project]
name = "gmail-mcp-server"
version = "0.1.0"
description = "Gmail MCP server for managing Gmail through Anthropic's Model Context Protocol"
requires-python = ">=3.9"
dependencies = [
    "mcp",
    "google-auth",
    "google-auth-oauthlib",
    "google-auth-httplib2",
    "google-api-python-client",
    "python-dotenv",
    "pydantic",
    "httpx",
]

[project.scripts]
gmail-mcp = "gmail_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gmail_mcp"]